+++ components/mentoring/MentoringListSection/MentoringList/index.tsx
@@ -1,31 +1,22 @@
 import styled from '@emotion/styled';
 import { useQuery } from '@tanstack/react-query';
 import Link from 'next/link';
-import ArrowDiagonalIcon from 'public/icons/icon-diagonal-arrow.svg';
 import { ReactNode, startTransition, useEffect, useState } from 'react';
 
 import { getMemberProfileById } from '@/api/endpoint_LEGACY/members';
 import Carousel from '@/components/common/Carousel';
 import useEventLogger from '@/components/eventLogger/hooks/useEventLogger';
 import { mentoringProvider } from '@/components/mentoring/data';
-import MentoringCard from '@/components/mentoring/MentoringCard';
-import { MENTOR_APPLICATION_URL, playgroundLink } from '@/constants/links';
-import { colors } from '@/styles/colors';
-import { textStyles } from '@/styles/typography';
-import { getScreenMaxWidthMediaQuery } from '@/utils';
+import MentoringCard from '@/components/mentoring/MentoringListSection/MentoringCard';
+import {
+  DESKTOP_LARGE_MEDIA_QUERY,
+  DESKTOP_SMALL_MEDIA_QUERY,
+  SCREEN_SIZE,
+  TABLET_MEDIA_QUERY,
+} from '@/components/mentoring/MentoringListSection/responsive';
+import { playgroundLink } from '@/constants/links';
 
 type ListType = 'carousel-large' | 'carousel-small' | 'scroll' | undefined;
-
-const SCREEN_SIZE = {
-  desktopLarge: { size: 1542, className: 'large-desktop-only' },
-  desktopSmall: { size: 1200, className: 'small-desktop-only' },
-  tablet: { size: 768, className: 'tablet-only' },
-  mobile: { size: 375, className: 'mobile-only' },
-};
-
-const DESKTOP_LARGE_MEDIA_QUERY = getScreenMaxWidthMediaQuery(`${SCREEN_SIZE.desktopLarge.size}px`);
-const DESKTOP_SMALL_MEDIA_QUERY = getScreenMaxWidthMediaQuery(`${SCREEN_SIZE.desktopSmall.size}px`);
-const TABLET_MEDIA_QUERY = getScreenMaxWidthMediaQuery(`${SCREEN_SIZE.tablet.size}px`);
 
 export default function MentoringList() {
   const [listType, setListType] = useState<ListType>();
@@ -56,10 +47,6 @@ export default function MentoringList() {
     moveCarousel: () => logClickEvent('mentoringCarouselButton'),
     clickCarouselCard: (mentorId: number) => logClickEvent('mentoringCard', { mentorId }),
     clickMentorApplicationButton: () => logClickEvent('mentorApplicationButton'),
-  };
-
-  const handleClickMentorApplicationButton = () => {
-    eventLogger.clickMentorApplicationButton();
   };
 
   const mentoringList = getMentoringList();
@@ -115,18 +102,6 @@ export default function MentoringList() {
 
   return (
     <Container>
-      <Header>
-        <Title>{`✨ NEW! \n아래의 멘토들이 \n멘티를 기다리고 있어요`}</Title>
-        <MentorApplicationButton
-          href={MENTOR_APPLICATION_URL}
-          target='_blank'
-          onClick={handleClickMentorApplicationButton}
-          rel='noopener'
-        >
-          멘토 등록을 하고싶다면?
-          <ArrowDiagonalIcon />
-        </MentorApplicationButton>
-      </Header>
       {(listType === undefined || listType === 'carousel-large') && (
         <StyledCarousel
           itemList={mentoringCardList}
@@ -155,14 +130,6 @@ export default function MentoringList() {
 }
 
 const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 40px;
-  align-items: center;
-  justify-content: center;
-  margin-top: 100px;
-  margin-bottom: 103px;
-
   .${SCREEN_SIZE.desktopSmall.className} {
     display: none;
   }
@@ -210,18 +177,6 @@ const Container = styled.div`
   }
 `;
 
-const Title = styled.div`
-  text-align: start;
-  line-height: 100%;
-  color: ${colors.white};
-
-  ${textStyles.SUIT_24_B}
-
-  @media ${TABLET_MEDIA_QUERY} {
-    white-space: pre-line;
-  }
-`;
-
 const StyledCarousel = styled(Carousel)`
   width: 1414px;
 
@@ -260,60 +215,5 @@ const MentoringScrollList = styled.div`
     & > .card:last-child {
       margin-right: 20px;
     }
-  }
-`;
-
-const MentorApplicationButton = styled.a`
-  display: flex;
-  gap: 6px;
-  align-items: center;
-  line-height: 100%;
-  color: ${colors.gray80};
-
-  ${textStyles.SUIT_18_M}
-
-  & > svg {
-    width: 20px;
-    height: 20px;
-
-    & > path {
-      fill: ${colors.gray80};
-    }
-  }
-
-  @media ${TABLET_MEDIA_QUERY} {
-    gap: 4px;
-    line-height: 15px;
-
-    ${textStyles.SUIT_12_M}
-
-    & > svg {
-      width: 12px;
-      height: 12px;
-    }
-  }
-`;
-
-const Header = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 1302px;
-
-  @media ${DESKTOP_LARGE_MEDIA_QUERY} {
-    width: 969px;
-  }
-
-  @media ${DESKTOP_SMALL_MEDIA_QUERY} {
-    flex-direction: column;
-    gap: 16px;
-    align-items: flex-start;
-    width: 636px;
-  }
-
-  @media ${TABLET_MEDIA_QUERY} {
-    gap: 12px;
-    padding: 0 20px;
-    width: 100%;
   }
 `;
