+++ components/mentoring/MentoringListSection/responsive.ts
@@ -1,0 +1,12 @@
+import { getScreenMaxWidthMediaQuery } from '@/utils';
+
+export const SCREEN_SIZE = {
+  desktopLarge: { size: 1542, className: 'large-desktop-only' },
+  desktopSmall: { size: 1200, className: 'small-desktop-only' },
+  tablet: { size: 768, className: 'tablet-only' },
+  mobile: { size: 375, className: 'mobile-only' },
+};
+
+export const DESKTOP_LARGE_MEDIA_QUERY = getScreenMaxWidthMediaQuery(`${SCREEN_SIZE.desktopLarge.size}px`);
+export const DESKTOP_SMALL_MEDIA_QUERY = getScreenMaxWidthMediaQuery(`${SCREEN_SIZE.desktopSmall.size}px`);
+export const TABLET_MEDIA_QUERY = getScreenMaxWidthMediaQuery(`${SCREEN_SIZE.tablet.size}px`);
